Moving DependencyScanningAction to its own header and source files.
